feat(all): 입력가격숫자범위를 long 자료형으로 허용함에따른 변경

+++ src/main/java/calculator/TotalProfitRateCalculator.java
@@ -12,21 +12,21 @@ public class TotalProfitRateCalculator {
     private static final int LOTTO_PRICE_UNIT = 1000;
 
 
-    public static double calculateTotalProfitRate(LottoWinningResult result, int buyQuantity) {
+    public static double calculateTotalProfitRate(LottoWinningResult result, long buyQuantity) {
         Map<LottoRank, Integer> lottoResult = result.getLottoResult();
-        int calculatedWinningAmount = calculateWinningAmount(lottoResult);
+        long calculatedWinningAmount = calculateWinningAmount(lottoResult);
         double totalProfitRate = ((PERCENTAGE_MULTIPLIER * calculatedWinningAmount) / (buyQuantity * LOTTO_PRICE_UNIT));
 
         return totalProfitRate;
     }
 
 
-    private static int calculateWinningAmount(Map<LottoRank, Integer> lottoResult) {
-        int totalWinningAmount = 0;
+    private static long calculateWinningAmount(Map<LottoRank, Integer> lottoResult) {
+        long totalWinningAmount = 0;
         Set<Entry<LottoRank, Integer>> resultEntries = lottoResult.entrySet();
         for (Entry<LottoRank, Integer> resultEntry : resultEntries) {
             LottoRank resultRank = resultEntry.getKey();
-            int winningAmountPrice = resultRank.getWinningAmount();
+            long winningAmountPrice = resultRank.getWinningAmount();
             totalWinningAmount = totalWinningAmount + winningAmountPrice * resultEntry.getValue();
         }
         return totalWinningAmount;
