+++ src/main/java/view/InputView.java
@@ -5,7 +5,7 @@ package view;
 public interface InputView {
 
 
-    int startView();
+    long startView();
 
 
     String inputWinningNumberView();
